NOISSUE Refactor; move stuff into standard ruby directories
