+++ bin/rubybot
@@ -1,0 +1,28 @@
+#!/usr/bin/env ruby
+
+require 'bundler/setup'
+require 'cinch'
+require 'rubybot/bot'
+require 'json'
+
+bot = Rubybot::Bot.new(
+    {
+        cfg_filename: 'config.json',
+        http_host:    '0.0.0.0',
+        http_port:    4567,
+        plugins:      [#BotAdmin,
+                       #Factoids,
+                       #Sed,
+                       #Tweet,
+                       #Github,
+                       #PingLists,
+                       #Youtube,
+                       #GithubLinker,
+                       #Cinch::HttpServer
+                      ]
+    })
+
+Signal.trap('TERM') { bot.quit }
+Signal.trap('HUP') { bot.reload_conf }
+
+bot.start
