+++ lib/rubybot/plugins/factoids/macro_semantic_error.rb
@@ -1,0 +1,13 @@
+module Rubybot
+  module Plugins
+    module Factoids
+      class MacroSemanticError < Exception
+        attr_accessor :msg
+
+        def initialize(msg = '')
+          @msg = msg
+        end
+      end
+    end
+  end
+end
