+++ lib/rubybot/plugins/http_server.rb
