+++ lib/rubybot/version.rb
@@ -1,0 +1,3 @@
+module Rubybot
+  VERSION = '1.0.0'
+end
