--- plugins/factoids/macro_semantic_error.rb
@@ -1,7 +1,0 @@
-class MacroSemanticError < Exception
-  attr_accessor :msg
-
-  def initialize(msg = '')
-    @msg = msg
-  end
-end
