--- plugins/github/Gemfile
@@ -1,1 +1,0 @@
-gem 'git.io', '~> 0.0.3'
