--- plugins/github_linker/Gemfile
@@ -1,2 +1,0 @@
-gem "octokit", "~> 3.8"
-gem "faraday-http-cache", "~> 1.1"
