--- plugins/google/Gemfile
@@ -1,2 +1,0 @@
-gem 'googleajax'
-gem 'sanitize'
