--- plugins/http_server/Gemfile
@@ -1,2 +1,0 @@
-gem 'sinatra'
-gem 'thin'
