--- plugins/tweet/Gemfile
@@ -1,1 +1,0 @@
-gem 'twitter'
