--- plugins/youtube/Gemfile
@@ -1,3 +1,0 @@
-gem 'google-api-client'
-gem 'chronic_duration'
-gem 'iso8601'
