--- src/elrosbot.rb
@@ -1,3 +1,0 @@
-require File.dirname(__FILE__) + '/elrosbot/acl'
-require File.dirname(__FILE__) + '/elrosbot/bot'
-require File.dirname(__FILE__) + '/elrosbot/plugin'
