--- src/elrosbot/plugin.rb
@@ -1,5 +1,0 @@
-module ElrosBot
-  module Plugin
-    include Cinch::Plugin
-  end
-end
